Moves us in a much simpler direction
This is now just a simple HTML page with some javascript that looks up a
users location and then populates most parts of the form. We're still
missing the lookup table for which region a state is in. Unless we can
generate it programmatically, the best we can do is to manually make a
csv file from the PDF provided by the state.